% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crea_dir.R
\name{crea_dir}
\alias{crea_dir}
\title{Création des dossiers de données}
\usage{
crea_dir(lb = list("test"))
}
\arguments{
\item{lb}{: liste des noms des dossiers}
}
\value{
No return value.
}
\description{
Fonction pour créer les dossiers contenant les données
}
