% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_pantheria_data.R
\name{dl_pantheria_data}
\alias{dl_pantheria_data}
\title{Download PanTHERIA dataset}
\usage{
dl_pantheria_data(overwrite = FALSE)
}
\arguments{
\item{overwrite}{a logical. If \code{TRUE}, the file will be downloaded again and
the previous version will be erased.}
}
\value{
No return value
}
\description{
This function downloads the PanTHERIA dataset (text file) hosted on the
GitHub repository \url{https://github.com/rdatatoolbox/datarepo/}.

PanTHERIA is a species-level database of life history, ecology, and geography of extant and recently extinct mammals. The dataset can be found in data/pantheria/PanTHERIA_1-0_WR05_Aug2008.txt. Metadata can be found here. Missing values are coded -999.
Citation: Kate EJ, Bielby J, Cardillo M et al. (2009) PanTHERIA: A species-level database of life history, ecology, and geography of extant and recently extinct mammals. Ecology, 90, 2648. DOI: 10.1890/08-1494.1

The file won't be downloaded if already exists locally (except if \code{overwrite = TRUE}).
The file \verb{PanTHERIA_1-0_WR05_Aug2008.txt} will be stored in \verb{data/pantheria/}. This folder will be created if required.
}
