% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_data}
\alias{download_data}
\title{Téléchargement des données}
\usage{
download_data(
  url = "https://raw.githubusercontent.com/rdatatoolbox/datarepo/main/data/pantheria/",
  file = list("test.txt"),
  destfolder = "test"
)
}
\arguments{
\item{url}{: url des données}

\item{file}{: liste des fichiers de données à télécharger}

\item{destfolder}{: nom du dossier final}
}
\value{
No return value.
}
\description{
Fonction pour télécharger les données dans les dossiers crées
}
